requirements.txt
REQUIREMENTS:
matplotlib~=3.1.3
numpy~=1.18.1
tqdm~=4.45.0
pandas~=1.0.3
seaborn~=0.10.0
pillow~=7.0.0
torchvision~=0.5.0
pytorch==1.4.0